+++ src/main/java/com/varshith/services/CustomerJpaDataAccessService.java
@@ -25,7 +25,7 @@ public class CustomerJpaDataAccessService implements CustomerDao {
     }
 
     @Override
-    public Optional<Customer> selectCustomerById(Integer id) {
+    public Optional<Customer> selectCustomerById(Long id) {
         return customerRepository.findById(id);
     }
 
@@ -40,7 +40,7 @@ public class CustomerJpaDataAccessService implements CustomerDao {
     }
 
     @Override
-    public void deleteCustomer(Integer id) {
+    public void deleteCustomer(Long id) {
             customerRepository.deleteById(id);
     }
 
@@ -50,7 +50,7 @@ public class CustomerJpaDataAccessService implements CustomerDao {
     }
 
     @Override
-    public boolean existsPersonWithId(Integer id) {
+    public boolean existsPersonWithId(Long id) {
         return customerRepository.existsCustomerById(id);
     }
 }
